Updated layout theme name for meeting 2
