+++ decks/node-beginner/meeting-2.mdx
@@ -1,7 +1,7 @@
 import { MinuteTimer } from '../components/minute-timer';
 export { theme } from './theme.js';
 import { Link } from '@reach/router';
-import CodeLayout from './code-layout-1';
+import LargerLayout from './larger-layout';
 import Split from './split';
 
 # [Digital Products Boosters](https://accenture.github.io/digital-products-boosters)
@@ -29,7 +29,7 @@ import Split from './split';
 
 ---
 
-<CodeLayout>
+<LargerLayout>
 
 ## Setup 🛠
 
@@ -50,7 +50,7 @@ To learn more about variables on Postman, go here:
 https://learning.getpostman.com/docs/postman/variables-and-environments/variables/#environments-in-postman
 </p>
 
-</CodeLayout>
+</LargerLayout>
 
 ---
 
@@ -103,14 +103,14 @@ GET https://api.github.com/users/:username
 
 ---
 
-<CodeLayout>
+<LargerLayout>
 
 # Exercise 3 - Answer
 
 - GET https://api.github.com/user/repos?affiliation=owner,collaborator
 - GET https://api.github.com/user/repos?affiliation=owner,collaborator&sort=created&direction=asc
 
-</CodeLayout>
+</LargerLayout>
 
 ---
 
@@ -142,7 +142,7 @@ Need help? https://developer.github.com/v3/repos/#enable-vulnerability-alerts
 
 ---
 
-<CodeLayout>
+<LargerLayout>
 
 # Exercise 5 - Answer
 
@@ -155,7 +155,7 @@ Need help? https://developer.github.com/v3/repos/#enable-vulnerability-alerts
 - Disable vulnerability alerts for a repo:
   - DELETE https://api.github.com/repos/:owner/:repo_name/vulnerability-alerts
 
-</CodeLayout>
+</LargerLayout>
 
 ---
